Clear LinkRecommendation suggestions on page save
The doSearchDataForIndex(2) hooks were originally used with the intention
to prevent an additional ElasticSearch index write when the article is
written and the ElasticSearch index is updated anyway.

However, since then the usage structure of these hooks and also the
implementation of the CirrusSearch indexing process changed a lot. Now,
the indexing process waits 10 minutes for any weighted_tags updates
and joins them together. That means we should just be able to
straight-forwardly delete our tags whenever a page is edited.

The new implementation always tries delete the search index tag without
checking whether it exists first. This is mainly done in order to
prevent dangling search index records from accumulating in case when the
database record had already been deleted otherwise. We do check for the
db-record existing in a replica in order to avoid unnecessary primary db
connections.

The PageSaveComplete hook is called in a PRESEND deferred update. So to
not delay the response to the user, we're now also deleting the search
index tag in a POSTSEND deferred update.

The new configuration option is introduced so that we can toggle this
change for individual wikis and observe its effects closely. Ultimately,
the goal is to enable it for all wikis and then remove the config
option.

See also: https://wikimedia.slack.com/archives/C0153LQ5G82/p1721838945390689

Related phab tasks: T347601, T372337, T314158, T364341

Bug: T364341
Bug: T372337
(cherry picked from commit 7d8a6210ecea676b1c89dac49b8d1e1fc34d2256)

+++ includes/NewcomerTasks/AddLink/LinkRecommendationHelper.php
@@ -12,6 +12,7 @@ use MediaWiki\Deferred\DeferredUpdates;
 use MediaWiki\Linker\LinkTarget;
 use MediaWiki\Page\ProperPageIdentity;
 use StatusValue;
+use Wikimedia\Rdbms\IDBAccessObject;
 
 /**
  * Shared functionality for various classes that consume link recommendations.
@@ -76,18 +77,29 @@ class LinkRecommendationHelper {
 	 * Delete recommendations for a given page from the database and optionally from the search index.
 	 * @param ProperPageIdentity $pageIdentity
 	 * @param bool $deleteFromSearchIndex
+	 * @param bool $allowJoiningSearchIndexDeletes if true, then CirrusSearch will join this with
+	 *                                             other events from the revision, waiting for up to 10 minutes
 	 */
 	public function deleteLinkRecommendation(
 		ProperPageIdentity $pageIdentity,
-		bool $deleteFromSearchIndex
+		bool $deleteFromSearchIndex,
+		bool $allowJoiningSearchIndexDeletes = false
 	): void {
-		DeferredUpdates::addCallableUpdate( function () use ( $pageIdentity ) {
-			$this->linkRecommendationStore->deleteByPageIds( [ $pageIdentity->getId() ] );
-		} );
+		if (
+			$this->linkRecommendationStore->getByPageId( $pageIdentity->getId(), IDBAccessObject::READ_NORMAL ) !== null
+		) {
+			DeferredUpdates::addCallableUpdate( function () use ( $pageIdentity ) {
+				$this->linkRecommendationStore->deleteByPageIds( [ $pageIdentity->getId() ] );
+			} );
+		}
 		if ( $deleteFromSearchIndex ) {
-			( $this->weightedTagsUpdaterProvider )()->resetWeightedTags(
-				$pageIdentity, [ LinkRecommendationTaskTypeHandler::WEIGHTED_TAG_PREFIX ]
-			);
+			DeferredUpdates::addCallableUpdate( function () use ( $pageIdentity, $allowJoiningSearchIndexDeletes ) {
+				( $this->weightedTagsUpdaterProvider )()->resetWeightedTags(
+					$pageIdentity,
+					[ LinkRecommendationTaskTypeHandler::WEIGHTED_TAG_PREFIX ],
+					$allowJoiningSearchIndexDeletes ? 'revision' : null
+				);
+			} );
 		}
 	}
